+++ NEWS.md
@@ -1,9 +1,19 @@
+# CFtime 1.7.2
+
+Maintenance release
+
+* Time units can use SI decimal multiples or sub-multiples, such as "µs" or "kiloyear".
+* Setting bad boundary values generates a warning instead of an error.
+* `CFCalendar` and subclasses use active fields for all properties.
+* CRAN landing page links to github.io documentation.
+
 # CFtime 1.7.1
 
 Maintenance release
 
 * Added `length` field to class `CFTime`.
 * Added `format.CFTime()` S3 method.
+* Added `"[.CFTime"` S3 method.
 * Documentation has been moved from vignettes to a website on github.io.
 
 # CFtime 1.7.0